% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assets.R
\name{get_assets}
\alias{get_assets}
\title{Get asset details}
\usage{
get_assets(asset = NULL)
}
\arguments{
\item{asset}{A string containing an asset symbol, e.g. \emph{btc}. Optional argument.}
}
\value{
data A list containing asset data.
}
\description{
Get asset details
}
