% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exchanges.R
\name{get_exchanges}
\alias{get_exchanges}
\title{Get exchange details}
\usage{
get_exchanges(exchange = NULL)
}
\arguments{
\item{exchange}{A string containing an exchange symbol, e.g. \emph{kraken}. Optional argument.}
}
\value{
data A list containing exchange data.
}
\description{
Get exchange details
}
