% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_markets.R
\name{get_markets}
\alias{get_markets}
\title{Get prices of cryptocurrencies}
\usage{
get_markets(route, pair = NULL, exchange = NULL, params = NULL)
}
\arguments{
\item{route}{A string containing a market endpoint. Possible values: \emph{price, prices, trades, summary, summaries, orderbook, orderbook/liquidity, orderbooks/calculator, ohlc} (required argument).}

\item{pair}{A string containing a pair symbol, e.g. \emph{btcusd} (optional argument).}

\item{exchange}{A string containing the exchange (optional argument). Run \code{get_exchanges()} to choose an exchange.}

\item{params}{A list containing query parameters. E.g., for the route \emph{ohlc}, this is \code{before}, \code{after} and \code{periods} (optional). See \emph{https://docs.cryptowat.ch/rest-api/markets} for further information.}
}
\value{
A list containing markets data.
}
\description{
Get prices of cryptocurrencies using the REST API of cryptowat.ch.
}
\references{
See \url{https://docs.cryptowat.ch/rest-api} for further information
}
\seealso{
\code{\link{markets}}, \code{\link{get_assets}}, \code{\link{get_exchanges}}, \code{\link{get_pairs}}
}
