% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_markets.R
\name{get_markets_as_df}
\alias{get_markets_as_df}
\title{Get data.frame with prices for cryptocurrencies}
\usage{
get_markets_as_df(pair, params = NULL, exchange = "kraken", route = "ohlc")
}
\arguments{
\item{pair}{A string containing a pair symbol, e.g. \emph{btcusd} (required argument).}

\item{params}{A list containing \code{before}, \code{after} and \code{periods} (optional). See \emph{https://docs.cryptowat.ch/rest-api/markets/ohlc} for further information.}

\item{exchange}{A string containing the exchange. Default is \emph{kraken}.}

\item{route}{A string containing market prices. Possible values: \emph{price, trades, summary, orderbook, ohlc}. Default is \emph{ohlc}.}
}
\value{
df.data A data.frame containing price data for given pair.
}
\description{
Get data.frame with prices for cryptocurrencies
}
