% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pairs.R
\name{get_pairs}
\alias{get_pairs}
\title{Get pair of currencies}
\usage{
get_pairs(pair = NULL)
}
\arguments{
\item{pair}{A string containing a pair symbol, e.g. \emph{btcusd}. Optional argument.}
}
\value{
data A list containing pair data.
}
\description{
Get pair of currencies
}
