% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markets.R
\name{markets}
\alias{markets}
\title{Get prices of cryptocurrencies}
\usage{
markets(pair, params = NULL, exchange = "kraken", datetime = TRUE,
        api_key = NULL, allowance = FALSE)
}
\arguments{
\item{pair}{A string containing a pair symbol, e.g. \emph{btcusd} (required argument). Run \code{get_pairs()} to find other available pairs.}

\item{params}{A list containing \code{before}, \code{after} and \code{periods} which is measured in seconds (optional). See \url{https://docs.cryptowat.ch/rest-api/markets/ohlc} for further information.}

\item{exchange}{A string containing the exchange. Default is \emph{kraken}. Run \code{get_exchanges()} to find other available exchanges.}

\item{datetime}{A logical. \code{TRUE} indicates that datetime type is used. \code{FALSE} indicates \emph{unix timestamp}. Default is \code{TRUE}.}

\item{api_key}{A string containing the API key. See \url{https://docs.cryptowat.ch/rest-api/rate-limit} to learn how to create an account and how to generate an API key.}

\item{allowance}{A logical (default is \code{FALSE}). If \code{TRUE} the function returns a list which includes allowance information, i.e. cost of the request, remaining credits and your account name.}
}
\value{
A data.frame containing prices of a given pair of currencies. If allowance is \code{TRUE}, \code{markets()} returns a list.
}
\description{
Get data.frame with prices of cryptocurrencies using the REST API of cryptowat.ch. The route is fixed to \emph{ohlc} and returns OHLC candlestick price data. The default is daily prices but can be changed with \code{params} and \code{periods}.
}
\examples{
\dontrun{
# Daily prices of Bitcoin in USD
df.btcusd <- markets("btcusd")
# Hourly prices of Bitcoin in USD for a specific time period
df.btcusd.hourly <- markets("btcusd", list(periods = 3600, before = 1609851600,
                                          after = 1609506000), datetime = FALSE)
# Daily prices of Bitcoin in Euro for a specific time period
df.btceur.daily <- markets("btceur", list(periods = 86400, before = "2021-05-12",
                                          after = "2021-01-01"), datetime = TRUE)
}

}
\references{
See \url{https://docs.cryptowat.ch/rest-api} for further information.
}
\seealso{
\code{\link{get_markets}}, \code{\link{get_assets}}, \code{\link{get_exchanges}}, \code{\link{get_pairs}}
}
