% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assets.R
\name{get_assets}
\alias{get_assets}
\title{Get asset details}
\usage{
get_assets(asset = NULL)
}
\arguments{
\item{asset}{A string containing an asset symbol, e.g. \emph{btc}. Optional argument.}
}
\value{
data A list or data.frame containing asset data.
}
\description{
Get asset information for cryptocurrencies.
}
\examples{
\dontrun{
asset.data <- get_assets("btc")
df.assets <- get_assets()
}

}
\references{
See \url{https://docs.cryptowat.ch/rest-api} for further information
}
\seealso{
\code{\link{get_markets}}, \code{\link{get_exchanges}}, \code{\link{get_pairs}}
}
