% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_markets.R
\name{get_markets}
\alias{get_markets}
\title{Get price data for crypto currencies}
\usage{
get_markets(pair, params, exchange = "kraken", route = "ohlc")
}
\arguments{
\item{pair}{A string containing a pair symbol, e.g. (btcusd). Required argument.}

\item{params}{A list containing before, after and periods (optional).}

\item{exchange}{A string containing the exchange. Default is kraken.}

\item{route}{A string containing market prices. Default is ohlc.}
}
\value{
df.markets A data frame containing price data for given arguments.
}
\description{
Get price data for crypto currencies
}
