% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pairs.R
\name{get_pairs}
\alias{get_pairs}
\title{Get pair of currencies}
\usage{
get_pairs(pair = NULL)
}
\arguments{
\item{pair}{A string containing a pair symbol, e.g. \emph{btcusd}. Optional argument.}
}
\value{
data A list or data.frame containing pair data.
}
\description{
Get details on pairs of (crypto)currencies.
}
\examples{
\dontrun{
btcusd.data <- get_pairs("btcusd")
df.pairs <- get_pairs()
}

}
\references{
See \url{https://docs.cryptowat.ch/rest-api} for further information
}
\seealso{
\code{\link{get_markets}}, \code{\link{get_assets}}, \code{\link{get_exchanges}}
}
